Update framework.py correction
The if was intended to avoid allegedly unnecessary updating of the Q-estimate. 
Nevertheless that updating in the case of new_q>=q only becomes unnecessary when the action was selected in a greedy manner, but it is selected in an epsilon-greedy manner, therefore the 'no updating' of the Q-estimate was preventing the agent from learning from random actions caused by the epsilon.